Remove unnecessary steps, add hints to set insert
Since git converts all files to UNIX end of line (from \r\n to \n),
we can assume that there will be no \r. If this is still the case,
std::getline should be able get a proper line based on the OS.

Moved branches insert outside of first iteration, so make use of the
already sorted dictionary, providing a hint to the insert.